requirements.txt
scrapetube
pytube==15.0.0
youtube-transcript-api
vertexai==1.43.0
google.cloud